requirements.txt
requests
responses
fastapi
openai==0.28
jinja2
python-multipart
python-dotenv
uvicorn
tavily-python
pdfminer.six
python-docx
